@@ -100,27 +100,28 @@
         "style-loader": "^0.23.1",
         "ts-loader": "^5.3.3",
         "ts-node": "^7.0.1",
-        "ts-node-dev": "^1.0.0-pre.44",
+        "ts-node-dev": "^1.0.0-pre.49",
         "tslint": "^5.20.1",
         "tslint-loader": "^3.6.0",
-        "typescript": "^3.7.5",
+        "typescript": "^3.9.5",
         "webpack": "^4.41.5",
-        "webpack-cli": "^3.3.10",
+        "webpack-cli": "^3.3.12",
         "webpack-dev-middleware": "^3.7.2",
-        "webpack-dev-server": "^3.10.1",
+        "webpack-dev-server": "^3.11.0",
         "webpack-hot-middleware": "^2.24.3"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.26",
-        "@fortawesome/free-regular-svg-icons": "^5.12.0",
-        "@fortawesome/free-solid-svg-icons": "^5.12.0",
-        "@fortawesome/react-fontawesome": "^0.1.8",
-        "@hig/flyout": "^1.2.0",
+        "@fortawesome/fontawesome-svg-core": "^1.2.29",
+        "@fortawesome/free-regular-svg-icons": "^5.13.1",
+        "@fortawesome/free-solid-svg-icons": "^5.13.1",
+        "@fortawesome/react-fontawesome": "^0.1.11",
+        "@hig/flyout": "^1.2.1",
         "@hig/theme-context": "^2.1.3",
-        "@hig/theme-data": "^2.13.0",
+        "@hig/theme-data": "^2.16.1",
+        "@types/google-maps": "^3.2.2",
         "@types/webscopeio__react-textarea-autocomplete": "^4.6.1",
-        "@webscopeio/react-textarea-autocomplete": "^4.6.3",
-        "adm-zip": "^0.4.13",
+        "@webscopeio/react-textarea-autocomplete": "^4.7.0",
+        "adm-zip": "^0.4.16",
         "archiver": "^3.1.1",
         "array-batcher": "^1.2.3",
         "async": "^2.6.2",
